Factor data flow functionality out into useDataFlow hook
